refactor(webui): plugin_catalog 拆分为多子模块包
将 956 行 plugin_catalog.py 拆为 classify/discovery/identity/metadata/role/visuals/version/catalog/config_module 子模块，经 __init__ 全量重导出保持导入面。子模块经包命名空间在调用时解析共享符号，保证测试字符串打桩传播。37 相关测试通过。
